task1 only

+++ task3 - disabled.js
@@ -2,8 +2,8 @@
 
 "use strict";
 
-var t1 = parseFloat(process.argv[2]),
-    t2 = parseFloat(process.argv[3]),
+var t1 = parseInt(process.argv[2]),
+    t2 = parseInt(process.argv[3]),
     hours = 0,
     minutes = 0,
     seconds = 0,
